+++ manager/manager/utils.py
@@ -22,7 +22,8 @@ import json
 import os
 import re
 import smtplib
-from datetime import timedelta
+import traceback
+from datetime import datetime, timedelta
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 from tempfile import TemporaryFile
@@ -41,6 +42,8 @@ from rest_framework.response import Response
 # Constants
 JSON_RESPONSE_LOCAL_STORAGE_NOT_ALLOWED = {"error": "Local storage not allowed."}
 JSON_RESPONSE_ERROR_NOT_VALID_JSON = {"error": "Not a valid JSON response."}
+
+ERROR_OBS_TICKETS = f"Error getting issues from {os.environ.get('JIRA_API_HOSTNAME')}"
 
 OBS_ISSUE_TYPE_ID = "10065"
 OBS_TIME_LOST_FIELD = "customfield_10106"
@@ -706,22 +709,29 @@ def get_jira_obs_report(request_data):
     response = requests.get(url, headers=headers)
     if response.status_code == 200:
         issues = response.json()["issues"]
-        return [
-            {
-                "key": issue["key"],
-                "summary": issue["fields"]["summary"],
-                "time_lost": (
-                    issue["fields"][OBS_TIME_LOST_FIELD]
-                    if issue["fields"][OBS_TIME_LOST_FIELD] is not None
-                    else 0.0
-                ),
-                "reporter": issue["fields"]["creator"]["displayName"],
-                "created": issue["fields"]["created"].split(".")[0],
-                "systems": parse_obs_issue_systems(issue),
-            }
-            for issue in issues
-        ]
-    raise Exception(f"Error getting issues from {os.environ.get('JIRA_API_HOSTNAME')}")
+        try:
+            return [
+                {
+                    "key": issue["key"],
+                    "summary": issue["fields"]["summary"],
+                    "time_lost": (
+                        issue["fields"][OBS_TIME_LOST_FIELD]
+                        if issue["fields"][OBS_TIME_LOST_FIELD] is not None
+                        else 0.0
+                    ),
+                    "reporter": issue["fields"]["creator"]["displayName"],
+                    "created": issue["fields"]["created"].split(".")[0],
+                    "systems": parse_obs_issue_systems(issue),
+                }
+                for issue in issues
+            ]
+        except KeyError as e:
+            traceback.print_exc()
+            raise Exception(
+                f"{ERROR_OBS_TICKETS}. "
+                f"Parsing JIRA response failed: missing field {e}"
+            )
+    raise Exception(ERROR_OBS_TICKETS)
 
 
 def get_client_ip(request):
@@ -1032,6 +1042,7 @@ def arrange_nightreport_email(report, plain=False):
         "https://rubinobs.atlassian.net/jira/software/c/projects/OBS/boards/232"
     )
     day_added = get_obsday_iso(report["day_obs"])
+    nightlydigest_urls = arrange_nightlydigest_urls_for_obsday(report["day_obs"])
 
     # TODO: Swap this hardcoded url by a dynamic one.
     # The service is meant to be run in the summit,
@@ -1039,6 +1050,7 @@ def arrange_nightreport_email(report, plain=False):
     # See: DM-43637
     url_rolex = f"https://summit-lsp.lsst.codes/rolex?log_date={day_added}"
 
+    NIGHTLYDIGEST_TITLE = "Nightly Digest:"
     SUMMARY_TITLE = "Summary:"
     WEATHER_TITLE = "Weather summary:"
     MAINTEL_SUMMARY_TITLE = "Simonyi summary:"
@@ -1053,7 +1065,10 @@ def arrange_nightreport_email(report, plain=False):
     CSCS_STATUS_TITLE = "CSCs status:"
 
     if plain:
-        plain_content = f"""{SUMMARY_TITLE}
+        return f"""{NIGHTLYDIGEST_TITLE}
+{nightlydigest_urls["simonyi"]}
+
+{SUMMARY_TITLE}
 {report["summary"]}
 
 {WEATHER_TITLE}
@@ -1082,7 +1097,6 @@ def arrange_nightreport_email(report, plain=False):
 
 {SIGNED_MSG}
 {", ".join(report["observers_crew"])}"""
-        return plain_content
 
     new_line_character = "\n"
     html_content = f"""
@@ -1107,6 +1121,11 @@ def arrange_nightreport_email(report, plain=False):
         </style>
     </head>
     <body>
+        <p>
+            {NIGHTLYDIGEST_TITLE}
+            <br>
+            <a href="{nightlydigest_urls["simonyi"]}">{nightlydigest_urls["simonyi"]}</a>
+        </p>
         <p>
             {SUMMARY_TITLE}
             <br>
@@ -1458,3 +1477,50 @@ def parse_cscs_status_to_plain_text(cscs_status):
         plain_text += f"{status}: {cscs_status[status]}\n"
 
     return plain_text.strip()
+
+
+def arrange_nightlydigest_urls_for_obsday(obsday):
+    """Arrange the URL for the nightly digest page
+    for a given observing day for both telescopes
+    Simonyi and AuxTel.
+
+    Parameters
+    ----------
+    obsday : `int`
+        The observing day in the format "YYYYMMDD"
+
+    Returns
+    -------
+    dict
+        Dictionary with the following keys
+        - simonyi: URL for the Simonyi telescope nightly digest page
+        - auxtel: URL for the AuxTel telescope nightly digest page
+
+    Raises
+    ------
+    ValueError
+        If obsday is not in the format "YYYYMMDD"
+    """
+
+    if len(str(obsday)) != 8:
+        raise ValueError(
+            f"Invalid obsday format: {obsday}. Expected format is 'YYYYMMDD'."
+        )
+
+    try:
+        datetime.strptime(str(obsday), "%Y%m%d")
+    except ValueError:
+        raise ValueError(
+            f"Invalid obsday format: {obsday}. Expected format is 'YYYYMMDD'."
+        )
+
+    return {
+        "simonyi": (
+            f"{settings.NIGHTLYDIGEST_BASE_URL}"
+            f"/?startDayobs={obsday}&endDayobs={obsday}&telescope=Simonyi"
+        ),
+        "auxtel": (
+            f"{settings.NIGHTLYDIGEST_BASE_URL}"
+            f"/?startDayobs={obsday}&endDayobs={obsday}&telescope=AuxTel"
+        ),
+    }